# api/openapi/foliofox-core-api.yaml
openapi: 3.0.3
info:
  title: FolioFox Core API
  description: |
    FolioFox is an eBook management and download automation system that provides comprehensive 
    APIs for managing book collections, automating downloads through various indexers, and 
    maintaining a high-quality digital library.
    
    ## Key Features
    - **Book Library Management**: Complete CRUD operations for books, authors, series, and metadata
    - **Download Automation**: Queue management with priority, retry logic, and status tracking
    - **Indexer Integration**: Support for multiple indexers (Prowlarr, Jackett, direct APIs)
    - **Search & Discovery**: Full-text search with advanced filtering and caching
    - **User Management**: Multi-user support with preferences and quality profiles
    - **Real-time Updates**: WebSocket support for live status updates
    
    ## Authentication
    All API endpoints require authentication via Bearer token unless otherwise specified.
    
    ## Rate Limiting
    - Standard endpoints: 1000 requests per hour per user
    - Search endpoints: 500 requests per hour per user
    - Download endpoints: 100 requests per hour per user
    
    ## Error Handling
    All errors follow RFC 7807 Problem Details format with consistent error codes and messages.
  version: 1.0.0
  contact:
    name: FolioFox API Support
    url: https://github.com/foliofox/foliofox
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
...
security:
  - bearerAuth: []
paths:
...
  /auth/login:
    post:
      tags: [Authentication]
      summary: User login
      description: Authenticate user and receive access token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                  example: admin
                password:
                  type: string
                  format: password
                  example: secure_password
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                    description: JWT access token
                  token_type:
                    type: string
                    enum: [Bearer]
                  expires_in:
                    type: integer
                    description: Token expiration time in seconds
                    example: 3600
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
  /auth/refresh:
    post:
      tags: [Authentication]
      summary: Refresh access token
      description: Get a new access token using refresh token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token:
                  type: string
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                  expires_in:
                    type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
  /users/profile:
    get:
      tags: [User Management]
      summary: Get current user profile
      responses:
        '200':
          description: User profile retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '401':
          $ref: '#/components/responses/Unauthorized'
    
    put:
      tags: [User Management]
      summary: Update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserProfileUpdate'
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /users/preferences:
    get:
      tags: [User Management]
      summary: Get user preferences
      responses:
        '200':
          description: User preferences retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPreferences'
        '401':
          $ref: '#/components/responses/Unauthorized'
    
    put:
      tags: [User Management]
      summary: Update user preferences
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserPreferencesUpdate'
      responses:
        '200':
          description: Preferences updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPreferences'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ===================================================================
  # BOOK LIBRARY MANAGEMENT
  # ===================================================================
  /books:
    get:
      tags: [Library Management]
      summary: List books in library
      description: |
        Retrieve books from the library with optional filtering, sorting, and pagination.
        Supports full-text search across titles, authors, descriptions, and tags.
      parameters:
        - name: page
          in: query
          description: Page number (1-based)
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Number of items per page
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: search
          in: query
          description: Full-text search query
          schema:
            type: string
            example: "science fiction"
        - name: author
          in: query
          description: Filter by author name
          schema:
            type: string
        - name: series
          in: query
          description: Filter by series name
          schema:
            type: string
        - name: genre
          in: query
          description: Filter by genre
          schema:
            type: string
        - name: language
          in: query
          description: Filter by language code
          schema:
            type: string
            example: "en"
        - name: format
          in: query
          description: Filter by available format
          schema:
            type: string
            enum: [epub, pdf, mobi, azw3, txt, djvu, fb2, rtf]
        - name: rating_min
          in: query
          description: Minimum rating filter
          schema:
            type: number
            minimum: 0
            maximum: 5
        - name: rating_max
          in: query
          description: Maximum rating filter
          schema:
            type: number
            minimum: 0
            maximum: 5
        - name: publication_year_min
          in: query
          description: Minimum publication year
          schema:
            type: integer
        - name: publication_year_max
          in: query
          description: Maximum publication year
          schema:
            type: integer
        - name: sort
          in: query
          description: Sort field
          schema:
            type: string
            enum: [title, author, publication_date, rating, created_at, updated_at]
            default: title
        - name: order
          in: query
          description: Sort order
          schema:
            type: string
            enum: [asc, desc]
            default: asc
        - name: include_files
          in: query
          description: Include file information in response
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Books retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  books:
                    type: array
                    items:
                      $ref: '#/components/schemas/Book'
                  pagination:
                    $ref: '#/components/schemas/PaginationInfo'
                  total_count:
                    type: integer
                    description: Total number of books matching filters
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Library Management]
      summary: Add book to library
      description: Add a new book with metadata to the library
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookCreate'
      responses:
        '201':
          description: Book created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: Book already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /books/{book_id}:
    get:
      tags: [Library Management]
      summary: Get book details
      description: Retrieve detailed information about a specific book
      parameters:
        - name: book_id
          in: path
          required: true
          description: Unique book identifier
          schema:
            type: integer
        - name: include_files
          in: query
          description: Include file information
          schema:
            type: boolean
            default: true
        - name: include_download_history
          in: query
          description: Include download history
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Book details retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookDetails'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Library Management]
      summary: Update book metadata
      description: Update book information and metadata
      parameters:
        - name: book_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookUpdate'
      responses:
        '200':
          description: Book updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Library Management]
      summary: Remove book from library
      description: |
        Remove book from library. This will also delete associated files if specified.
        **Warning**: This operation cannot be undone.
      parameters:
        - name: book_id
          in: path
          required: true
          schema:
            type: integer
        - name: delete_files
          in: query
          description: Also delete associated files from disk
          schema:
            type: boolean
            default: false
      responses:
        '204':
          description: Book deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Cannot delete - book has active downloads
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /books/{book_id}/files:
    get:
      tags: [Library Management]
      summary: Get book files
      description: List all files associated with a book
      parameters:
        - name: book_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Book files retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  files:
                    type: array
                    items:
                      $ref: '#/components/schemas/BookFile'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    post:
      tags: [Library Management]
      summary: Add file to book
      description: Associate a new file with an existing book
      parameters:
        - name: book_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, format]
              properties:
                file:
                  type: string
                  format: binary
                  description: Book file to upload
                format:
                  type: string
                  enum: [epub, pdf, mobi, azw3, txt, djvu, fb2, rtf]
                quality_score:
                  type: integer
                  minimum: 0
                  maximum: 100
                  default: 50
                is_primary:
                  type: boolean
                  default: false
                  description: Set as primary format for this book
      responses:
        '201':
          description: File added successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookFile'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '413':
          description: File too large
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /books/{book_id}/files/{file_id}:
    get:
      tags: [Library Management]
      summary: Download book file
      description: Download a specific book file
      parameters:
        - name: book_id
          in: path
          required: true
          schema:
            type: integer
        - name: file_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: File download
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Library Management]
      summary: Delete book file
      description: Remove a specific file from a book
      parameters:
        - name: book_id
          in: path
          required: true
          schema:
            type: integer
        - name: file_id
          in: path
          required: true
          schema:
            type: integer
        - name: delete_from_disk
          in: query
          description: Also delete file from disk
          schema:
            type: boolean
            default: true
      responses:
        '204':
          description: File deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ===================================================================
  # SEARCH & DISCOVERY
  # ===================================================================
  /search:
    get:
      tags: [Search & Discovery]
      summary: Search books across indexers
      description: |
        Search for books across configured indexers with advanced filtering options.
        Results are cached for improved performance on repeated queries.
      parameters:
        - name: query
          in: query
          required: true
          description: Search query
          schema:
            type: string
            example: "Foundation Isaac Asimov"
        - name: indexers
          in: query
          description: Comma-separated list of indexer IDs to search
          schema:
            type: string
            example: "1,2,3"
        - name: formats
          in: query
          description: Comma-separated list of preferred formats
          schema:
            type: string
            example: "epub,pdf"
        - name: languages
          in: query
          description: Comma-separated list of language codes
          schema:
            type: string
            example: "en,es"
        - name: min_quality
          in: query
          description: Minimum quality score
          schema:
            type: integer
            minimum: 0
            maximum: 100
        - name: max_size_mb
          in: query
          description: Maximum file size in MB
          schema:
            type: integer
            minimum: 1
        - name: timeout
          in: query
          description: Search timeout in seconds
          schema:
            type: integer
            minimum: 5
            maximum: 60
            default: 30
        - name: limit
          in: query
          description: Maximum number of results per indexer
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
        - name: use_cache
          in: query
          description: Use cached results if available
          schema:
            type: boolean
            default: true
      responses:
        '200':
          description: Search completed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  query:
                    type: string
                    description: Original search query
                  results:
                    type: array
                    items:
                      $ref: '#/components/schemas/SearchResult'
                  total_results:
                    type: integer
                  indexers_searched:
                    type: array
                    items:
                      type: object
                      properties:
                        indexer_id:
                          type: integer
                        indexer_name:
                          type: string
                        result_count:
                          type: integer
                        response_time_ms:
                          type: integer
                        error:
                          type: string
                          nullable: true
                  search_duration_ms:
                    type: integer
                  cached:
                    type: boolean
                    description: Whether results came from cache
                  cache_expires_at:
                    type: string
                    format: date-time
                    nullable: true
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
  /search/suggestions:
    get:
      tags: [Search & Discovery]
      summary: Get search suggestions
      description: Get autocomplete suggestions based on library content
      parameters:
        - name: query
          in: query
          required: true
          description: Partial search query
          schema:
            type: string
            minLength: 2
        - name: type
          in: query
          description: Type of suggestions
          schema:
            type: string
            enum: [all, title, author, series, genre]
            default: all
        - name: limit
          in: query
          description: Maximum number of suggestions
          schema:
            type: integer
            minimum: 1
            maximum: 20
            default: 10
      responses:
        '200':
          description: Suggestions retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  suggestions:
                    type: array
                    items:
                      type: object
                      properties:
                        text:
                          type: string
                        type:
                          type: string
                          enum: [title, author, series, genre]
                        count:
                          type: integer
                          description: Number of matching items in library
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /search/history:
    get:
      tags: [Search & Discovery]
      summary: Get search history
      description: Retrieve user's search history with statistics
      parameters:
        - name: limit
          in: query
          description: Number of recent searches to return
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: days
          in: query
          description: Number of days to look back
          schema:
            type: integer
            minimum: 1
            maximum: 365
            default: 30
      responses:
        '200':
          description: Search history retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  history:
                    type: array
                    items:
                      $ref: '#/components/schemas/SearchHistoryEntry'
                  total_searches:
                    type: integer
                  unique_queries:
                    type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'

    delete:
      tags: [Search & Discovery]
      summary: Clear search history
      description: Clear user's search history
      parameters:
        - name: days
          in: query
          description: Clear history older than N days (if not specified, clears all)
          schema:
            type: integer
            minimum: 1
      responses:
        '204':
          description: Search history cleared
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ===================================================================
  # DOWNLOAD MANAGEMENT
  # ===================================================================
...
  /downloads/queue/{download_id}:
    get:
      tags: [Download Management]
      summary: Get download details
      description: Get detailed information about a specific download
      parameters:
        - name: download_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Download details retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DownloadDetails'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Download Management]
      summary: Update download
      description: Update download properties like priority or destination
      parameters:
        - name: download_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DownloadUpdate'
      responses:
        '200':
          description: Download updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DownloadQueueItem'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Cannot update - download in progress
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags: [Download Management]
      summary: Cancel/remove download
      description: |
        Cancel a download if it's in progress, or remove it from the queue.
        Completed downloads can only be removed from the queue, not deleted from disk.
      parameters:
        - name: download_id
          in: path
          required: true
          schema:
            type: integer
        - name: delete_partial
          in: query
          description: Delete partial download file
          schema:
            type: boolean
            default: true
      responses:
        '204':
          description: Download cancelled/removed
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
          example: 1
        username:
          type: string
          example: admin
        email:
          type: string
          format: email
          example: admin@foliofox.local
          nullable: true
        is_active:
          type: boolean
          example: true
        is_admin:
          type: boolean
          example: false
        last_login:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...